TODO scan cleanup; enable csrf; Response message in invalid request; File type check in image upload;  Refactor structure
