--- src/main/java/com/prc391/patra/config/security/MethodSecurityConfig.java
@@ -1,4 +1,0 @@
-package com.prc391.patra.config.security;
-
-public class MethodSecurityConfig {
-}
