--- src/main/java/com/prc391/patra/config/security/PatraLogoutHandler.java
@@ -1,4 +1,0 @@
-package com.prc391.patra.config.security;
-
-public class PatraLogoutHandler {
-}
