--- src/main/java/com/prc391/patra/config/security/permissionevaluator/TaskSecurityExpressionHandler.java
@@ -1,4 +1,0 @@
-package com.prc391.patra.config.security.permissionevaluator;
-
-public class TaskSecurityExpressionHandler {
-}
