--- src/main/java/com/prc391/patra/config/security/permissionevaluator/TaskSecurityExpressionRoot.java
@@ -1,4 +1,0 @@
-package com.prc391.patra.config.security.permissionevaluator;
-
-public class TaskSecurityExpressionRoot {
-}
