+++ src/main/java/com/prc391/patra/security/PatraLogoutHandler.java
@@ -1,0 +1,4 @@
+package com.prc391.patra.security;
+
+public class PatraLogoutHandler {
+}
