+++ src/main/java/com/prc391/patra/sheets/SheetService.java
@@ -1,6 +1,6 @@
 package com.prc391.patra.sheets;
 
-import com.prc391.patra.config.security.PatraUserPrincipal;
+import com.prc391.patra.security.PatraUserPrincipal;
 import com.prc391.patra.constant.SecurityConstants;
 import com.prc391.patra.exceptions.EntityExistedException;
 import com.prc391.patra.exceptions.EntityNotFoundException;
