+++ src/main/java/com/prc391/patra/utils/BaseResponse.java
@@ -1,0 +1,11 @@
+package com.prc391.patra.utils;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class BaseResponse {
+    private boolean success;
+    private String message;
+}
